adapt all code base to new hash map lookup

+++ src/on_tree.c
@@ -392,7 +392,7 @@ on_tree_tooltip(GtkWidget *w, int32 x, int32 y, gboolean k, GtkTooltip *t, void 
                 for (HardLink *link = hard_links; link; link = link->next) {
                     int32 n;
 
-                    n = snprintf(tip_buffer + offset, SIZEOF(tip_buffer) - offset - 5,
+                    n = snprintf(tip_buffer + offset, (size_t)(SIZEOF(tip_buffer) - offset - 5),
                                  "\n%s%s", RSYNC_HARDLINK, link->name);
                     offset += n;
                     if (offset >= (SIZEOF(tip_buffer) - 5)) {
